Changed ISO string for dates to use ' ' separator between date and time rather than 'T' separator.
If an instance was deleted on Nov 7th, this would cause a date range of Nov 6 to Nov 7 to include the hours used on Nov 7th, when it is supposed to stop at Nov 7th midnight. This is because OpenStack stores its date strings as ' ' separated, while Javascript defaults to inputting 'T' separated dates, and mysql compares dates using string comparison, so 2016-11-07T00:00:00 (midnight Nov 7) would be considered greater than 2016-11-07 17:00:00 (5PM Nov 7). Which is wrong.